+++ src/app/(pages)/admin/page.jsx
@@ -1,5 +1,6 @@
 "use client";
 import StateCard from "@/components/admin/adminHome/stateCard";
+import UserManagement from "@/components/admin/adminHome/userManagement";
 import DataLoader from "@/components/shared/dataLoader/dataLoader";
 import { useSession } from "next-auth/react";
 import { FiClipboard, FiUser, FiUserCheck, FiUserPlus, FiUsers } from "react-icons/fi";
@@ -23,30 +24,7 @@ const AdminDashboard = () => {
 
       {/* User Management Section */}
       <div>
-        <h2 className="text-3xl font-bold mb-4 text-[#5C0096]">User Management</h2>
-        <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6 mb-12 text-white">
-          <div className="bg-[#bb00a2] p-6 rounded-lg shadow-lg flex items-center">
-            <FiUserCheck className="text-4xl text-white mr-4" />
-            <div>
-              <h2 className="text-xl font-bold">Total Active Users</h2>
-              <p className="text-3xl font-semibold">120</p>
-            </div>
-          </div>
-          <div className="bg-[#17A2B8] p-6 rounded-lg shadow-lg flex items-center">
-            <FiUserPlus className="text-4xl text-white mr-4" />
-            <div>
-              <h2 className="text-xl font-bold">New Users <span className="text-sm font-normal">{"(This Week)"}</span></h2>
-              <p className="text-3xl font-semibold">10</p>
-            </div>
-          </div>
-          <div className="bg-[#7700ff] p-6 rounded-lg shadow-lg flex items-center">
-            <FiUser className="text-4xl text-white mr-4" />
-            <div>
-              <h2 className="text-xl font-bold">Top Quiz Attempts</h2>
-              <p className="text-3xl font-semibold flex items-center gap-2">15 <span className="text-sm font-normal">{"(John Doe)"}</span></p>
-            </div>
-          </div>
-        </div>
+        <UserManagement />
       </div>
 
       {/* Recent Quizzes Section */}
